@@ -228,10 +228,10 @@ func testLoad(t *testing.T, cache Cache, conf config.AllConfigs) {
 	numOps := 512
 	dal.SetOpRetryProps(5, 1000, 1000)
 
-	rand.Seed(time.Now().Unix())
+	someRand := rand.New(rand.NewSource(time.Now().Unix()))
 	for i := 0; i < numOps; i++ {
 		go func(ch chan bool, id int) {
-			apiKeyNum := rand.Intn(int(testdbs.HopsworksAPIKey_ADDITIONAL_KEYS))
+			apiKeyNum := someRand.Intn(int(testdbs.HopsworksAPIKey_ADDITIONAL_KEYS))
 			apiKey := fmt.Sprintf("%016d.%s", apiKeyNum, testdbs.HopsworksAPIKey_SECRET)
 			DB := testdbs.DB001
 			err := cache.ValidateAPIKey(&apiKey, &DB)
@@ -364,11 +364,11 @@ func testLoadWithBadKeys(t *testing.T, cache Cache, conf config.AllConfigs) {
 
 	ch := make(chan bool)
 
-	rand.Seed(time.Now().Unix())
+	someRand := rand.New(rand.NewSource(time.Now().Unix()))
 	numOps := 1000
 	for i := 0; i < numOps; i++ {
 		go func(ch chan bool) {
-			badAPIKey := badKeys[rand.Intn(len(badKeys))]
+			badAPIKey := badKeys[someRand.Intn(len(badKeys))]
 			DB := testdbs.DB001
 			err := cache.ValidateAPIKey(&badAPIKey, &DB)
 			if err != nil {
